+++ src/js/apiService.js
@@ -20,9 +20,6 @@ export default {
     this.page = currentPage;
 
     this.currentRequest = load('currentRequest');
-    console.log(load('currentRequest'));
-    console.log(this.currentRequest);
-
     switch (this.currentRequest) {
       case HOME:
         this.urlApi = `https://api.themoviedb.org/3/trending/movie/week?api_key=${this.keyApi}&page=${this.page}&language=${this.language}`;
