+++ src/js/refs.js
@@ -24,11 +24,9 @@ const refs = {
 
 //  =======
 
-  // pageNavRef: () => document.querySelector('.site-nav__list'),
-  // pageHomeRef: () => document.querySelector('[data-request="home"]'),
+
   pageLibraryRef: () => document.querySelector('[data-request="library"]'),
-  // pageWatchedRef: () => document.querySelector('[data-request="watched"]'),
-  // pageQueueRef: () => document.querySelector('[data-request="queue"]'),
+
 
   //=======
     divLightbox: document.querySelector('.lightbox'),
